Extract receipt functions from main.rs into receipts.rs module
Move persist_receipt(), persist_transition_receipt(), show_receipts(),
and ReceiptEntry struct into a dedicated receipts.rs module to reduce
main.rs from 2405 to 2200 lines. Update .line-ratchet.toml ceiling
from 2406 to 2201.
